Combine XAlloc.[ch] into XUtils.[ch]
