+++ FunctionBar.c
@@ -8,7 +8,6 @@ in the source distribution for its full text.
 #include "FunctionBar.h"
 #include "CRT.h"
 #include "RichString.h"
-#include "XAlloc.h"
 #include "XUtils.h"
 
 #include <assert.h>
