+++ Hashtable.c
@@ -6,7 +6,7 @@ in the source distribution for its full text.
 */
 
 #include "Hashtable.h"
-#include "XAlloc.h"
+#include "XUtils.h"
 
 #include <stdlib.h>
 #include <assert.h>
